# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 家庭 AI 学习助手 API
  version: 0.6.0
  description: Self-hosted Household account/session and learning contract vertical slices.
servers:
  - url: http://localhost:8000
paths:
...
  /households/{household_id}/sessions/{session_id}/captures:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - $ref: '#/components/parameters/SessionId'
    get:
      operationId: listCaptures
      summary: List Capture metadata visible to the bound child in a study session
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Capture metadata. This endpoint never returns media bytes or object URLs.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Capture'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createCapture
      summary: Register bounded image metadata and require manual correction
      description: No image bytes, signed URLs, object keys, or OCR Provider calls are accepted in this contract version.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaptureRequest'
      responses:
        '201':
          description: Capture metadata registered with needs_correction status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        '200':
          description: Idempotent Capture-create replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /households/{household_id}/captures/{capture_id}/corrections:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: correctCapture
      summary: Append a manual correction for a Capture owned by the bound child
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectCaptureRequest'
      responses:
        '201':
          description: Manual correction appended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureCorrection'
        '200':
          description: Idempotent correction replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureCorrection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Capture version or idempotency conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/save:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: saveCapture
      summary: Save a Capture image beyond its default retention window
      description: Parent-only local/CI operation. Saving does not expose media bytes, object keys, or storage credentials.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Capture image is marked parent-saved.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /households/{household_id}/captures/{capture_id}/media:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteCaptureMedia
      summary: Immediately delete one Capture image
      description: Parent-only local/CI operation. Deletion is private-object scoped, retryable, and does not delete Capture metadata.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Capture image deletion completed or was replayed.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '503':
          description: Object storage deletion is temporarily unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/sessions/{session_id}/capture-uploads:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - $ref: '#/components/parameters/SessionId'
    post:
      operationId: beginCaptureUpload
      summary: Create a pending Capture and issue one short-lived private PUT URL
      description: The response has no separate object-key field. The signed URL is short-lived and must not be logged or persisted by clients.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCaptureRequest'
      responses:
        '201':
          description: Pending Capture created and a short-lived upload URL issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureUpload'
        '200':
          description: Idempotent replay while the Capture upload remains pending. A replacement short-lived URL may be issued.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaptureUpload'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
        '503':
          description: Private object storage is temporarily unavailable; retry with the same idempotency key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/upload-confirmations:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: confirmCaptureUpload
      summary: Confirm a server-verified private object before manual correction
      description: The API verifies object media type and byte size server-side. It never accepts an object key or bytes from the client.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmCaptureUploadRequest'
      responses:
        '201':
          description: Private object verified; Capture advances to needs_correction.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        '200':
          description: Idempotent upload-confirmation replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capture'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Object verification, Capture state/version, or idempotency conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/ocr-jobs:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      operationId: enqueueOcrJob
      summary: Enqueue local OCR for a server-confirmed Capture
      description: The endpoint only records a bounded job request. OCR runs in a separate local Worker and returns manually confirmable candidates; it never accepts image bytes, object keys, or external Provider selection.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnqueueOcrJobRequest'
      responses:
        '202':
          description: OCR job accepted by the local/CI queue.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrJobReceipt'
        '200':
          description: Idempotent OCR job enqueue replay.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrJobReceipt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Capture is not ready for OCR or the idempotency key conflicts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/captures/{capture_id}/ocr-jobs/{job_id}:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: capture_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getOcrJob
      summary: Read bounded local OCR job status
      description: The response contains stable lifecycle fields only; Provider errors, image bytes, object keys, and OCR text are not returned by this endpoint.
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: OCR job status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OcrJobReceipt'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    HouseholdId:
      name: household_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  securitySchemes:
    BearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque 256-bit session token
      description: Flutter session token; only a SHA-256 digest is stored by the API.
    SessionCookie:
      type: apiKey
      in: cookie
      name: study_session
      description: HttpOnly SameSite=Lax Web session cookie.
  schemas:
...
    CreateCaptureRequest:
      type: object
      additionalProperties: false
      required: [media_type, byte_size, content_sha256]
      properties:
        media_type:
          type: string
          enum: [image/jpeg, image/png]
        byte_size:
          type: integer
          minimum: 1
          maximum: 8000000
        content_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
    Capture:
      type: object
      additionalProperties: false
      required: [id, household_id, child_id, session_id, media_type, byte_size, content_sha256, status, version, created_at]
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        media_type:
          type: string
          enum: [image/jpeg, image/png]
        byte_size:
          type: integer
        content_sha256:
          type: string
          pattern: '^[a-f0-9]{64}$'
        status:
          type: string
          enum: [upload_pending, needs_correction, corrected]
        version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
    CaptureUpload:
      type: object
      additionalProperties: false
      required: [capture, upload_url, upload_expires_at]
      properties:
        capture:
          $ref: '#/components/schemas/Capture'
        upload_url:
          type: string
          format: uri
          description: Short-lived PUT URL; clients must not persist or log it.
        upload_expires_at:
          type: string
          format: date-time
    ConfirmCaptureUploadRequest:
      type: object
      additionalProperties: false
      required: [expected_capture_version]
      properties:
        expected_capture_version:
          type: integer
          minimum: 1
...
    CorrectCaptureRequest:
      type: object
      additionalProperties: false
      required: [expected_capture_version, corrected_text]
      properties:
        expected_capture_version:
          type: integer
          minimum: 1
        corrected_text:
          type: string
          minLength: 1
          maxLength: 1000
...
    CaptureCorrection:
      type: object
      additionalProperties: false
      required: [id, capture_id, household_id, child_id, sequence, corrected_text, created_at]
      properties:
        id:
          type: string
          format: uuid
        capture_id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        child_id:
          type: string
          format: uuid
        sequence:
          type: integer
          minimum: 1
        corrected_text:
          type: string
        created_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid synthetic principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Principal role cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource is not visible in the caller's household.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdempotencyConflict:
      description: Idempotency key was reused with a different payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...